+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   via option `--only-deployed`
 - New empty solver option that simply makes sure that all SMT queries return
   unknown
+- Allow `verifyInputs` to return partial expressions
 
 ## Fixed
 - We now extract more Keccak computations than before from the Props to assert